Initial commit
### **Step 0: Starting Point**

This is the starting point for the final project, which is provided to you in the [course repository](https://github.com/udacity/ud867/tree/master/FinalProject). It contains an activity with a banner ad and a button that purports to tell a joke, but actually just complains. The banner ad was set up following the instructions here:

[https://developers.google.com/mobile-ads-sdk/docs/admob/android/quick-start](https://developers.google.com/mobile-ads-sdk/docs/admob/android/quick-start)

You may need to download the Google Repository from the Extras section of the Android SDK Manager.

You will also notice a folder called backend in the starter code. It will be used in step 3 below, and you do not need to worry about it for now.

When you can build an deploy this starter code to an emulator, you're ready to move on.